labels

+++ src/ostd.c
@@ -41,8 +41,8 @@ s8 s8_from(AllocFn alloc, const char* s) {
 }
 
 const char* s8_to_c(AllocFn alloc, s8 s) {
-  char* n = alloc((1 + s.len) * sizeof(char));
-  memcpy(n, s.str, s.len * sizeof(char));
+  char* n = alloc((u64)(1 + s.len) * sizeof(char));
+  memcpy(n, s.str, (u64)s.len * sizeof(char));
   n[s.len] = '\0';
   return (const char*)n;
 }
